Single Mapper
Changing ImageFilter class to use the Hadoop framework.
This simply translates the same main logic to a single Mapper.
The single Mapper will read a whole image file and then apply the filter to it.
Custom InputFormat/RecordReader make sure that input is not split and that entire file is read by the mapper.
No Reducer is used yet, so the Mapper's output is the final output.